+++ pytorch/src/main/resources/org/bytedeco/pytorch/presets/torch_include.h
@@ -1483,6 +1483,7 @@
 #include "torch/csrc/distributed/c10d/ProcessGroup.hpp"
 #include "torch/csrc/distributed/c10d/comm.hpp"
 #include "torch/csrc/distributed/c10d/default_comm_hooks.hpp"
+//#include "torch/csrc/distributed/c10d/symm_mem/intra_node_comm.hpp"
 #include "c10/util/ApproximateClock.h"
 #include "torch/csrc/distributed/c10d/reducer_timer.hpp"
 // #include "torch/csrc/autograd/functions/basic_ops.h" // Not on Windows