+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Enable distributed package using NCCL in presets for PyTorch ([issue #1402](https://github.com/bytedeco/javacpp-presets/issues/1402))
  * Fix adapter for `c10::intrusive_ptr` in presets for PyTorch ([issue #1678](https://github.com/bytedeco/javacpp-presets/issues/1678))
  * Include `FileStore.hpp`, `TCPStore.hpp`, and `gloo/transport/tcp/device.h` header files in presets for PyTorch ([discussion #1712](https://github.com/bytedeco/javacpp-presets/discussions/1712))
  * Introduce `linux-arm64` builds for ONNX ([pull #1697](https://github.com/bytedeco/javacpp-presets/pull/1697))